Fixed unintended debouncing, updated ingredients endpoint
