+++ src/components/SearchBarTagsSelector/SearchBarTagsSelector.scss
