+++ src/pages/home/Home.tsx
@@ -4,7 +4,8 @@ import React, { ReactElement, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import './Home.scss';
 import { AxiosResponse } from 'axios';
-import SearchBar from '../../components/SearchBar/SearchBar';
+import { useSelector } from 'react-redux';
+import AutocompleteSearchBar from '../../components/AutocompleteSearchBar/AutocompleteSearchBar';
 import FilledButton from '../../components/FilledButton/FilledButton';
 import { ROUTES } from '../../routes/routes';
 import {
@@ -15,8 +16,15 @@ import Tag from '../../components/Tag/Tag';
 import { ProductCriteria, productCriteriaToUrl } from '../../services/product.service';
 import { TagColor } from '../../theme/tagColor';
 import { ObjectWithNameAndId } from '../../types/objectWithNameAndId';
+import SearchBar from '../../components/SearchBar/SearchBar';
+import { RootState } from '../../store/reducers';
+
+const MAX_INGREDIENTS_SUGGESTIONS = 50;
 
 const Home = (): ReactElement => {
+  const hasAllergens: boolean = useSelector((state: RootState) => state.like.dislikedIngredients)
+    ?.length > 0 ?? false;
+
   const navigate = useNavigate();
   // todo: keep ingredients in a provider to not duplicate code between Home and Products list
   const [phrase, setPhrase] = useState<string>('');
@@ -41,7 +49,7 @@ const Home = (): ReactElement => {
       setIngredientsSuggestions(null);
       return;
     }
-    getIngredientsApi(query, 50).then(
+    getIngredientsApi(query, MAX_INGREDIENTS_SUGGESTIONS, hasAllergens).then(
       (value: AxiosResponse<IngredientObject[]>) => {
         setIngredientsSuggestions(value.data);
       },
@@ -83,7 +91,7 @@ const Home = (): ReactElement => {
             placeholder="e.g. shampoo"
             onChange={(value) => setPhrase(value)}
           />
-          <SearchBar
+          <AutocompleteSearchBar
             id="ingredient-search"
             label="Ingredients"
             placeholder="e.g. shea butter"
